Major update - automation/improvement
No longer need to select good lines by hand and paste them in - safeguards I've implemented do this automatically.

Majorly cleaned up the code - put everything in a function, added a few comments, tried to improve some of the naming. It still needs cleaned up further though

Now it knows when to stop - the program will terminate once it's written a song that has a length of around 1/3rd of the lyrics you give it to draw from.